requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow

django-cors-headers==4.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1

psycopg2-binary>=2.9.9,<3.0
psycopg-binary==3.2.9
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0

selenium>=4.16.0,<5.0
webdriver-manager>=4.0.1,<5.0
